+++ .devcontainer/dev/setup-remote-vm.sh
@@ -9,7 +9,7 @@ set -eo pipefail
 BUILDARCH=$([ "$(uname -i)" = "aarch64" ] && echo "arm64" || echo "amd64")
 HOSTNAME="knfsd-dev-ec2"
 USERNAME="ubuntu"
-VERSION="1.1.0-alpha.16"
+VERSION="1.1.0-alpha.17"
 
 ## set env vars for build env only
 export HOME=/home/${USERNAME}
@@ -142,29 +142,29 @@ KNFSD_BATS_CORE_VERSION=1.13.0
 # https://github.com/psf/black/releases
 KNFSD_BLACK_VERSION=25.11.0
 # https://github.com/boto/boto3/tags
-KNFSD_BOTO3_VERSION=1.40.72
+KNFSD_BOTO3_VERSION=1.42.0
 # https://hub.docker.com/r/bridgecrew/checkov/tags
-KNFSD_CHECKOV_VERSION=3.2.493
+KNFSD_CHECKOV_VERSION=3.2.495
 # https://github.com/codespell-project/codespell/releases
 KNFSD_CODESPELL_VERSION=2.4.1
 # https://github.com/editorconfig-checker/editorconfig-checker/releases
-KNFSD_EDITORCONFIG_VERSION=3.4.1
+KNFSD_EDITORCONFIG_VERSION=3.6.0
 # https://github.com/golangci/golangci-lint/releases
-KNFSD_GOLANGCI_LINT_VERSION=2.6.1
+KNFSD_GOLANGCI_LINT_VERSION=2.6.2
 # https://go.dev/dl/
 KNFSD_GOLANG_VERSION=1.25.4
 # https://github.com/python/mypy/tags
-KNFSD_MYPY_VERSION=1.18.2
+KNFSD_MYPY_VERSION=1.19.0
 # https://github.com/hashicorp/packer/releases
-KNFSD_PACKER_VERSION=1.14.2
+KNFSD_PACKER_VERSION=1.14.3
 # https://github.com/pre-commit/pre-commit/releases
-KNFSD_PRECOMMIT_VERSION=4.4.0
+KNFSD_PRECOMMIT_VERSION=4.5.0
 # https://pypi.org/project/psycopg/
-KNFSD_PSYCOPG_VERSION=3.2.12
+KNFSD_PSYCOPG_VERSION=3.3.0
 # https://github.com/pylint-dev/pylint/tags
-KNFSD_PYLINT_VERSION=4.0.2
+KNFSD_PYLINT_VERSION=4.0.4
 # https://github.com/semgrep/semgrep/releases
-KNFSD_SEMGREP_VERSION=1.143.0
+KNFSD_SEMGREP_VERSION=1.144.0
 # https://pypi.org/project/shellcheck-py/
 KNFSD_SHELLCHECK_PY_VERSION=0.11.0.1
 # https://github.com/mvdan/sh/releases
@@ -172,9 +172,9 @@ KNFSD_SHFMT_VERSION=3.12.0
 # https://github.com/hashicorp/terraform/releases
 KNFSD_TERRAFORM_VERSION=1.2.9
 # https://github.com/gruntwork-io/terragrunt/releases
-KNFSD_TERRAGRUNT_VERSION=0.93.8
+KNFSD_TERRAGRUNT_VERSION=0.93.12
 # https://github.com/terraform-linters/tflint/releases
-KNFSD_TFLINT_VERSION=0.59.1
+KNFSD_TFLINT_VERSION=0.60.0
 # https://github.com/aquasecurity/tfsec/releases
 KNFSD_TFSEC_VERSION=1.28.14
 # https://github.com/aquasecurity/trivy/releases