+++ deployment/terraform-module-knfsd/modules/loadbalancer/variables.tf
@@ -8,10 +8,10 @@ variable "VERSION" {
   description = "(Required) The version of the KNFSD File Cache."
   type        = string
   nullable    = false
-  default     = "1.1.0-alpha.16"
+  default     = "1.1.0-alpha.17"
   validation {
     condition     = can(regex("^(?P<major>0|[1-9]\\d*)\\.(?P<minor>0|[1-9]\\d*)\\.(?P<patch>0|[1-9]\\d*)(?:-(?P<prerelease>(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+(?P<buildmetadata>[0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$", var.VERSION))
-    error_message = "VERSION must be a valid semantic version 2.0.0 format. Example: \"1.1.0-alpha.16\"."
+    error_message = "VERSION must be a valid semantic version 2.0.0 format. Example: \"1.1.0-alpha.17\"."
   }
 }
 
@@ -94,6 +94,12 @@ variable "NFS_PORTS" {
       name       = "nfs-callback"
     }
   }
+}
+
+variable "VPC_CIDR" {
+  description = "(Required) List of CIDR blocks to allow in security group rules."
+  type        = list(string)
+  nullable    = false
 }
 
 variable "LOADBALANCER_IP" {